Moved PosZInt.scala code into PosInts.scala.
